Refactored file utilities
