+++ api/src/utils/constants.py
@@ -1,0 +1,6 @@
+# Add default paths here so it can be used by multiple files
+# Not in .env because the default should be committed
+
+# FROM ROOT DIR
+input_dir = 'data/input'
+output_dir = 'data/output'
